fix(main): allow server to run with either a js defined schema or a gql langauge schema
