+++ src/main.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import { graphqlHTTP } from "express-graphql";
-import { schema } from "./schema-sdl";
-import { roots } from "./roots";
+import { schema } from "./schema-js";
+//import { roots } from "./roots";
 
 const app = express();
 
@@ -9,7 +9,7 @@ app.use(
   "/graphql",
   graphqlHTTP({
     schema: schema,
-    rootValue: roots,
+    //rootValue: roots,
     graphiql: true,
   })
 );
